Stability update
Fixed some bugs relating to selecting connections with the incorrect
index and when a connection is closed while interacting with it.
Before, if a client closed a connection while the CLI was being
accessed it would require the user to type `quit` in order to close the
client shell.  The client shell now closes if realizes that the
connection is closed so the Oyster shell can be interacted with right
away.