Added server and shell plugins!
When a server is connected to a client, the commands can be intercepted
using plugins before being sent.  Commands can also be added to the
Oyster shell using the shell plugins.  They follow the same rules as
the client plugins except the server passes it’s `Server` instance to
the `Plugin.run` method instead of a `Client` instance, along with the
extra data it got.