Plugins now track whether a `Plugin` is `enabled` using the `enabled` attribute.  Update to all plugins that list plugins as well.
Plugins must now include an `enabled` attribute to dictate whether or
not they are enabled or disabled.  This is mostly for debugging, but
can be a nice way of stopping a plugin from working without removing
it.

All of the plugins that list plugins now retrieve them using the
get_*_plugins() methods so more information can be displayed.