Huge update that adds support for different connection types, which allows for more functionality(progress bars for downloads, etc).  The Oyster shell and Client shell can now be scripted.  More in description
Connection Types update:

HeaderServer, TerminatingServer, HeaderClient, TerminatingClient

A HeaderServer communicates with a HeaderClient, and TerminatingServer
communicates with a TerminatingServer.  Basically, the way connections
send and receive data can be customized.  These connection types show
how that works and are located in the `connection.py` module.

Before, a termination string had to be sent in order to tell the client
or server that it was done transmitting data.  The
HeaderServer/HeaderClient connection type sends a header with the
amount of bytes that will be sent after the header information.  This
allows things like progress bars on downloads to be written(the `get`
command will display a progress bar for downloads if using the
HeaderServer/HeaderClient).  Some other plugins may need to be updated,
but I will get them updated here soon(I think the upload plugin will
need to be updated for the HeaderClient/Server).

Scripting Oyster Shell and Client Shell:

Plugins will be written that make it super simple to read in commands
from a file, however I haven’t written them yet.  The
`server.Server.start_client_shell` and `server.Server.open_oyster`
methods have a keyword argument called `commands` that you can give a
list of commands to.  Instead of opening up a prompt for a user, it
will just run the commands, then return.