Added ability to use a text editor using the `#` command.  Default editor is suplemon. More in description
This plugin lets you use a text editor on you local machine to edit a
file located on the client machine.  The file data from the client is
sent to the server and saved as a temporary file, then `Popen` is used
to open the temporary file with a text editor.  The contents of the
temporary file are sent back to the client once the text editor is
closed, so if the file was changed and saved, the client will save
those changes!

Syntax is:

To edit a file: `# {filepath}`
To use a specific editor: `# {editor} {filepath}`

Example:

    `<127.0.0.1> Some/Dir/Oyster> # client.py`
    `<127.0.0.1> Some/Dir/Oyster> # nano client.py`