Patched so files besides text files can be edited.  Tested commands in description
Tested with image magick running

`magick drink.jpg -resize 50% drink.jpg`

The caveat being that the program must be the first arg, and the
filename __must__ be the last arg.

The image is transmitted from the client to the server and saved in a
temporary file.  The temporary file is modified server-side and the
contents of the new file are transmitted back to the client which
overwrites the original file resulting in an image that is 50% smaller.