Added more documentation and refactored the `Editor` plugin to use tagged filenames for more complicated editing.
It used to be that the `Editor` plugin would only work if the command
was structured like:

`{program} {flags} {filename}`

Now you just have to tag filenames with `<f>` so that they can be
handled correctly.

Example of using imagemagick to resize a picture:

    `/Users/user/Pictures> magick ~/Pictures/drink.jpg -resize 50%
drink.jpg`

  becomes

    `/Users/user/Pictures> # magick <f>~/Pictures/drink.jpg -resize 50%
<f>drink.jpg`

If you ran the first example, it would try to use imagemagick on the
client machine whereas the second example would grab the filedata(for
the first tagged filename in the command), from the client machine and
then use imagemagick to edit it locally before sending the modified
data back to the client.